requirements.txt
# Core
anthropic>=0.42.0
python-dotenv>=1.0.0

# Data & Indicators
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Async
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# Exchange (Phase 10)
# ccxt>=4.0.0

# Dashboard
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Telegram (Phase 9)
# python-telegram-bot>=21.0

# News (Phase 2)
# feedparser>=6.0.0
